refactor(nlp): engines interfaces moved
