+++ packages/botonic-nlp/src/preprocess/engines/es/stemmer-es.ts
@@ -1,4 +1,4 @@
-import { Stemmer } from '../../types'
+import { Stemmer } from '../types'
 
 export class StemmerEs implements Stemmer {
   readonly locale = 'es'
